Changes to ImageMap/ImageReduce

+++ lab03/src/mrf/example/ImageReduce.java
@@ -4,11 +4,11 @@ import java.util.ArrayList;
 
 import mrf.tasks.ReduceCallable;
 
-public class ImageReduce implements ReduceCallable<ArrayList<Byte>> {
+public class ImageReduce implements ReduceCallable<ArrayList<Integer>> {
 
 	@Override
-	public ArrayList<Byte> reduce(ArrayList<ArrayList<Byte>> f) {
-		ArrayList<Byte> res = new ArrayList<Byte>();
+	public ArrayList<Integer> reduce(ArrayList<ArrayList<Integer>> f) {
+		ArrayList<Integer> res = new ArrayList<Integer>();
 		for(int i = 0; i < f.size(); i++){
 			res.addAll(f.get(i));
 		}
